requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
PyMuPDF==1.24.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.3.0
google-genai==0.4.0
python-dotenv==1.0.1
pydantic==2.7.0
